+++ src/js/elections/components/Ridings.jsx
@@ -12,12 +12,19 @@ import css from './ridings.scss';
 
 export default class Ridings extends React.Component {
   render() {
+    const numberOfRidings = this.props.allRidings ? this.props.allRidings.length : null;
+
     const ridings = this.props.allRidings
       ? this.props.allRidings.map(riding => (
           <article key={`art_${riding.id}`} className={css.riding} id={`art_${riding.id}`}>
             {
               <Fragment>
-                <RidingHeader ridingName={riding.name} key={`heading_${riding.id}`} />
+                <RidingHeader
+                  ridingName={riding.name}
+                  key={`heading_${riding.id}`}
+                  ridingId={riding.id}
+                  numRidings={numberOfRidings}
+                />
                 <ErrorBoundary>
                   <ol className={css.riding__list}>
                     <Bar ridingResults={riding.results} key={riding.id} />